requirements.txt
alembic==0.9.9
autopep8==1.3.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
coverage==4.5.1
flasgger==0.8.1
Flask==0.12.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
httpauth==0.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
mistune==0.8.3
nose==1.3.7
psycopg2==2.7.4
pyasn1==0.4.2
pycodestyle==2.3.1
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.7.0
python-editor==1.0.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
urllib3==1.22
Werkzeug==0.14.1
yapf==0.20.2
